Code refactor with CountryNamesPlusMore.
